RAG/OpenAI/requirements.txt
streamlit
faiss-cpu
numpy
langchain-openai
langchain-community
pypdf2
pymupdf